Chore: Refactor usage of `current_profile`
`current_profile` can't be reliably used for ChA-mentor-combo accounts.
I refactored the chapter assignment functionality to explicitly use a
certain profile type.

For the interrupt screen, only students and mentors will see this and
there aren't any current mentor-student-combo accounts, so we'll use
either the mentor profile (if it exists) or the student one.

When admins create a new chapter assignment, it will always prioritize
the ChA's chapter assignment; the button on the front-end reflects that
it's for the ChA too.

Refs: #5075