Bug: Update onboarding status for off-platform legal docs
This will address an issue where the chapter and chapter ambassador's
onboarding statuses weren't getting updated for an off-platform legal
agreement. The onboarding check was using a stale copy of the legal
agreements. Using `reload` will refresh the legal agreements to make
sure the onboarding check is using the updated version.

Refs: #5057