UI: Update how ChA legal agreement status is displayed
There are some cases on Preview where a chapter ambassador's chapter
affiliation agreement doesn't have a status which is causing an error.
I'm not sure how this is happening (it could just be old, corrupt
testing data). Anywho, for the time being, this update will make it so
their status won't be displayed if there isn't one.

Refs: #5108