Feature: Display clubs on Chapter selection page
The main changes here:
- A new club selector service that returns clubs in an account's state/province and
  country
- Updates to the chapter selection page to accommodate clubs
  - This page could use some clean up, but for now I'm just focusing on
    getting it working

There will need to be some text changes to makes things clearer that now
someone can select a chapter or club. And saving a club won't work right
now, the back-end still needs to be refactored to allow for clubs to be
saved to the database.

Refs: #5200