Feature: Chapterable groundwork
This will introduce the concept of something being chapterable, i.e.
something chapter-like, something that can either be a chapter or a
club; this will allow one bucket for chapterable assignments that can be
used for either chapters or clubs (or any future chapter-like thing).

The work here just lays the groundwork for chapterables.

Some examples of how chapterables will be used:

- a chapterable assignment would be for a club or a chapter
- a chapterable view partial would be able to handle a chapter or a club
- a student's chapterable would be their chapter or club
- a chapter ambassador's chapterables would be their chapters
- a club ambassador's chapterables would be their clubs

A definition of -able:
https://www.oxfordlearnersdictionaries.com/definition/english/able_2

From the dictionary definitions:

- that can or must be
- having the quality of

I think of clubs as mini-chapters, they "have the quality of" being a
chapter and "can be" a chapter.

Refs: #5187