Feature: Refactor chapter assignments to allow for clubs to be saved
The main change here is to embrace the chapterable concept. All things
(or pretty much everything) that can refer to a chapter or a club has
been renamed to `chapterable`. This includes:

- `force_chapter_selection`
- `chapter_account_assignments`
  - and the corresponding associations and related pieces

The changes here also include the ability to save clubs. In order to
save a club (or a chapter), the chapterable id and type (either Chapter
or Club) must be passed to the controller; to do this, the radio button
values now contain both the id and type, then the controller splits up
the values to create the chapterable.

Refs: #5187