Chore: Chapterable cleanup work
Renaming and fixing things that were missed during the chapterable
refactoring.

I also added a `present?` guard around assigning mentors to chapters
when they join a student's team b/c sometimes it seems a student on a
team can be in a state where they are not assigned to a chapter yet, so
it would try to assign the mentor to the student's `NullChapter`, not it
won't do this.

Refs: #5187