UI: Update how clubs header is displayed
This should address the issue of the "Club" header being displayed
twice.

Now if there aren't any clubs in someone's state/province, but there are
clubs in someone's country, the "Club" header will be displayed along
with the "Show All Clubs in my country" button (instead of displaying
the header after the button was clicked).

Refs: #5200