Chore: Send separate parent/guardian first and last names to Salesforce
This will split up the parent/guardian name into two parts before
sending it to Salesforce, one part (the first part separated by
whitespace) will be sent as the first name, and the remaining part will
be sent as the last name.

And then in Salesforce the first and last name can be combined to form
the full name like it is on the platform.

Refs: #5265