Chore: Handle Morocco location
This follows our existing approach for addressing Palestine, except this
is for Morocco. I just copied the Palestine functionality and updated it
for Morocco and it seems to be working. We'll need to test further to
see if this has any unintended side effects.

Best as I can tell, this intercepts the response from the geocoder
provider and creates a custom location object.

Refs: #5241