Chore: Send updated student profile info to Salesforce
When upserting contact info in Salesforce, students have additional
info that needs to be sent (students are the only profile type that has
additional contact info). The additional info is:

- parent guardian email address
- parent guardian full name
- parent guardian first name
- parent guardian last name
- DOB

But in order to send this additional info, the `profile_type` needs to
be specified when calling `upsert_contact` and it wasn't. This means
that parent/guardian info and DOB for students wasn't getting updated
after it was initially created.

This change will address updating parent guardian contact info.

Refs: #5265