Chore: Send updated student DOB to Salesforce
There was an issue where if a student's DOB was updated it wouldn't get
updated in Salesforce. This is because all participants have basic info
that is sent to Salesforce, but students have extra info that needs to
be sent to Salesforce, one piece of additional info is their DOB.

In order to send their updated DOB to Salesforce `profile_type` needs to
be specified when making a call to `upsert_contact_info` and it wasn't,
so that's what the change here is doing.

Refs: #5285