Chore: Rename is_a_club_ambassador? -> club_ambassador?
Part of the refactoring work I did was to rename `is_a_club_ambassador?`
to `club_ambassador?`, this code was added after the refactoring work I
did, so it needed to get updated.

I also updated `is_an_ambassador?` to `chapter_ambassador?`. The
`is_an_ambassador?` functionality is confusing now because it **only**
refers to chapter ambassadors, it does **not** include club ambassadors.

Refs: #5270