Chore: Address Salesforce duplicate participant records
When the Salesforce service is called to setup a new account for the
season (`#setup_account_for_current_season`), a program participant
record (in Salesforce) gets created for them.

This functionality was intended to be called and used only once, and
that is when a participant is registered to the current season.

There is an issue though because this functionality can be called more
than once, like when a chapter or club ambassador profile gets added to
a mentor account. And this would create two program participant records
in Salesforce.

To fix this, instead of always creating the program participant record
(when we're setting up a new account in Salesforce), we'll first check
if one exists and update that record instead of creating a new one, and
if a record doesn't exits then we'll create it.

Refs: #5255