Fix: Unaffilated list for club ambassadors not assigned to a club
It's possible for a club ambassador to not be assigned to a club. If this happens they are still able to access the Club Admin Activity. The unaffiliated datagrid will error out if they're not assigned to a club (or chapter).

```
NoMethodError: undefined method `country_code' for #<NullChapterable:0x00007fd115046990>
```

This update will address this issue.

Refs: #5371