Chore: Add rake task to remove students from the unaffiliated list
More details are in #5396, but basically we ran the rake task to assign
students to a chapter, but they weren't removed from the unaffiliated
list; and that's because the original chapter assignment rake task was
meant for "limbo" students who aren't on the unaffiliated list.

This rake task will take list of email addresses (like the original
chapter assignment rake task) and if they're a student it will remove
them from the unaffiliated list.

Refs: #5396