Chore: Ensure one primary chapter assignment per profile type per season
This will add a validation to ensure that only primary chapterable
assignment per profile type per season is created.

It's possible right now that a mentor would have a non-primary
assignment first (by joining a team; we have an issue to make that their
primary assignment, #5364, but until that gets addressed), we need to
accommodate this scenario.

Also, a lot of the specs had to be updated because the specs were
creating more than one assignment and only one primary assignment is
valid now.

Refs: #5130