Fix: Error when unauthenticated ambassadors try to access pages
If an ambassador tries to access certain pages without logging in, it
will cause an error.

Trying to access this page:

- localhost:3000/chapter_ambassador/participants

It will cause an error `NoMethodError: undefined method 'chapter_ambassador?' for nil:NilClass`

The changes here should address that.

I removed the custom error message, "You must be signed in as XYZ to go
there!" because we don't want unnecessary details being leaked in our
error messages.

Refs: #5416