Chore: Add service to remove invalid mentor chapterable assignments
This new service will remove invalid mentor chapterable
assignments.

A chapterable assignment would be invalid when for example they leave a
team (that is in a different chapter or club, and they're not mentoring
another team that belongs to the chapter or club), in this case that
chapterable assignment should be removed.

Refs: #5431