Chore: Add judging location to Salesforce API calls
This will add "judging location" when updating a judge's "program
participant" record in Salesforce.

If a judge is attending an RPE the value will be "RPE (In-person)"
and if not (if they are removed from an event) it will be "Platform
(Virtual)".

From our dev team conversations, we just want to be able to identify RPE
judges and this should let us do that; by default everyone is a virtual
judge.

Refs: #4891