Fix: ChA trying to export judge list from an RPE
If a chapter ambassador tries to export judges from an RPE and there are
judges who were invited but not registered to the platform yet, it will
cause an error. This is because invited judges are a `UserInvitation`
record, not a `JudgeProfile` record. The error is being cause because
there are two methods that are being called that only existed on the
`JudgeProfile` model and not on `UserInvitation`:

- `#assigned_teams_for_event`
- `#company_name`

I added these^ methods to `UserInvitation` and it seems to work now (I
can't export the CSV locally, but I was getting an error and then made
the fix and it seems to work now).

Refs: #5575