Chore: Restrict club ambassadors from accessing judge mode
This is a quick fix to make the
`ENABLE_CHAPTER_AMBASSADOR_SWITCH_TO_JUDGE` config setting apply to both
chapter ambassadors and club ambassadors, so that way we can prevent
club ambassadors from judging. In the future we can rename this config
setting or split it into two, one for chapter ambassadors and one for
club ambassadors.

Refs: #5581