Chore: Unset "force chapterable selection" flag
When a new season starts, students and mentors will get reassigned to
their chapterable and be shown the chapterable selection screen (via the
`force_chapterable_selection` boolean flag).

This new functionality will make it so that when a student
or mentor confirms their chapterable, we will unset the
`force_chapterable_selection` flag in the database (preventing the
chapterable selection screen from showing up again) and take them to
their dashboard.

Refs: #5694