Feature: Tracking who made a chapterable assignment
This will implement basic tracking for who made a chapterable
assignment. Right now a chapterable assignment can be associated to an
account or a team.

When an admin, ambassador, mentor or student makes a chapterable
assignment we will track the account that made the assignment.

When a mentor joins a team we will track which team made the assignment
(we'll have to infer that the assignment was made b/c they joined a
team).

Refs: #5645