Chore: Remove JS test suite and supporting code
This will remove our JS test suite and all the supporting code:

The JS test suite doesn't provide a lot of value, we don't work on it
(add new tests or update any existing ones), it only tests Vue.js stuff
which we don't work on very often and will be replacing soon.

We should probably remove the test suite after we've replaced Vue.js,
but we're in a clean-up mode and the tests aren't providing a lot of
value anyway if they're not catching any bugs for us.

Refs: #5919