Feature: Allow admins and ambassadors to create teams
This will give admins and ambassadors the ability to create a team for a
student or mentor.

This will add a new section under the "Help them with their team
matching" section in the admin area that has a text field (for the team
name) and a submit button. After providing a team name and submitting
the form, it will create the team and assign the student or mentor to
that team.

I extracted the "Help them with their team matching"
content/functionality into a new partial.

I added a custom success message for admins or ambassadors creating a
team, otherwise it would use the default, "Your team has been created!
Lookin' good!" message.

Refs: #5859