Chore: Update specs to work with latest version of Chrome
Using the latest version of Chrome, a bunch of our tests were failing.
This will address most of them.

Here is one of the errors:

```
Selenium::WebDriver::Error::UnknownError:
unknown error: unhandled inspector error: {"code":-32000,"message":"Node with given id does not belong to the document"}
(Session info: chrome=141.0.7390.122)
````

To solve this error, I used this trick by setting an expectation on the
current path:

- https://github.com/SeleniumHQ/selenium/issues/15401#issuecomment-2726498973

For other errors, I added more expectations about the content being
displayed on the page, and in a few I added the "sleep 1" hack to keep
things moving along.

This seems to fix the majority of our specs and makes CircleCI pass the
majority of the time.

Ultimately, this will move us away from having to rely on a pinned
Chrome version, and make the tests more reliable on CircleCI.

Refs: #5884