Chore: Update parental consent status check links to fix surveys
During the Rails 7 update, the "check parental consent status" link
broke, when I was fixing that, I broke the surveys. The "check
parental consent status" link won't always be on the page, when it
wasn't on the page it was causing a JS error, which was causing the
surveys not to load.

The updates here will:

- Add a conditional for the "check parental consent status" link
- Update the survey specs

Refs: #5884