Chore: Fix double-redirect call when bulk-adding teams to an RPE
This will move the redirect when bulk-adding teams to an RPE to outside
of the block; if there was more than one chunk (of 200 records) it would
call redirect multiple times. Moving the redirect outside of the block
will make sure all of the chunks get processed and then it will
redirect.

Refs: #5912