Feature: Allow documents more than 120 days to be voided
This will make it so that when documents are more than 120 days old they
can be marked as voided on our platform by an admin.

Before this change if an admin tried to void one of these documents it
would not be voided on our platform b/c voiding on our platform requires
a successful "void" response from Docusign, but they weren't getting
voided in Docusign b/c after 120 days the documents expire and are
automatically marked as voided, and this would not return a successful
response and hence they wouldn't get marked as voided on our platform.
The change here will mark documents as voided on our platform if the
document is more than 120 days old.

> By default, envelopes expire 120 days after sending. When an envelope
> expires, the status changes to Voided and it can no longer be viewed or
> signed by recipients.

https://support.docusign.com/s/document-item?language=en_US#d10061e238&bundleId=gbo1643332197980&topicId=kyt1578456367748.html&_LANG=enus

Refs: #5989